+++ raidionicsrads/Processing/tumor_features_computation.py
@@ -3,7 +3,7 @@ import traceback
 import numpy as np
 from typing import Tuple, List
 from scipy.ndimage.measurements import center_of_mass
-from scipy.ndimage import measurements
+from scipy.ndimage import measurements, label
 from skimage.measure import regionprops
 from scipy.ndimage import binary_closing, _ni_support
 from scipy.ndimage import distance_transform_edt, binary_erosion, generate_binary_structure
@@ -30,7 +30,7 @@ def compute_volume(volume: np.ndarray, spacing: tuple) -> float:
         volume_pixels = np.count_nonzero(volume)
         volume_mmcube = voxel_size * volume_pixels
         volume_ml = volume_mmcube * 1e-3
-        result = np.round(volume_ml, 2)
+        result = float(round(volume_ml, 2))
     except Exception as e:
         raise ValueError('Volume computation failed with: {}'.format(e))
 
@@ -58,7 +58,7 @@ def compute_shape(volume: np.ndarray, spacing: tuple) -> Tuple[float, float, flo
             feret = c.feret_diameter_max * np.prod(spacing[0:3])
             equivalent = c.equivalent_diameter_area * np.prod(spacing[0:3])
 
-    return long_axis, short_axis, feret, equivalent
+    return float(long_axis), float(short_axis), float(feret), float(equivalent)
 
 
 def compute_multifocality(volume: np.ndarray, spacing: tuple,
@@ -127,7 +127,7 @@ def compute_multifocality(volume: np.ndarray, spacing: tuple,
     except Exception as e:
         raise ValueError('Multifocality computation failed with: {}'.format(e))
 
-    return multifocal_status, multifocal_elements, multifocal_largest_minimum_distance
+    return multifocal_status, multifocal_elements, float(multifocal_largest_minimum_distance)
 
 
 def compute_lateralisation(volume: np.ndarray, brain_mask: np.ndarray,
@@ -183,15 +183,15 @@ def compute_lateralisation(volume: np.ndarray, brain_mask: np.ndarray,
             right_side_percentage = np.count_nonzero((brain_mask == 1) & (volume != 0)) / np.count_nonzero((volume != 0))
             left_side_percentage = np.count_nonzero((brain_mask == 2) & (volume != 0)) / np.count_nonzero((volume != 0))
 
-            left_hemisphere_percentage = np.round(left_side_percentage * 100., 2)
-            right_hemisphere_percentage = np.round(right_side_percentage * 100., 2)
+            left_hemisphere_percentage = round(left_side_percentage * 100., 2)
+            right_hemisphere_percentage = round(right_side_percentage * 100., 2)
             midline_crossing = True if max(left_hemisphere_percentage, right_hemisphere_percentage) < 100. else False
         else:
             pass
     except Exception as e:
         raise ValueError('Lateralisation computation failed with {}'.format(e))
 
-    return left_hemisphere_percentage, right_hemisphere_percentage, midline_crossing
+    return float(left_hemisphere_percentage), float(right_hemisphere_percentage), midline_crossing
 
 
 def compute_resectability_index(volume: np.ndarray, resectability_map: np.ndarray) -> Tuple[float, float, float]:
@@ -228,7 +228,7 @@ def compute_resectability_index(volume: np.ndarray, resectability_map: np.ndarra
         avg_resectability = total_resectability / tumor_voxels_count
     except Exception as e:
         raise ValueError('Resectability index computation failed with {}'.format(e))
-    return residual_tumor_volume, resectable_volume, avg_resectability
+    return float(residual_tumor_volume), float(resectable_volume), float(avg_resectability)
 
 
 def compute_hd95(reference, result, voxelspacing=None, connectivity=1):
